requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
requests==2.32.3